Shulker Tendon variable delay

+++ src/main/java/uwu/juni/recharged/content/blocks/ShulkerTendonBlock.java
@@ -10,20 +10,28 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.util.RandomSource;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.DirectionalBlock;
+import net.minecraft.world.level.block.EntityBlock;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition.Builder;
+import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import uwu.juni.recharged.Recharged;
+import uwu.juni.recharged.content.blocks.block_entities.ShulkerTendonBlockEntity;
 
 @ParametersAreNonnullByDefault
-public class ShulkerTendonBlock extends DirectionalBlock {
-	public static final int DELAY = 4;
+public class ShulkerTendonBlock extends DirectionalBlock implements EntityBlock {
+	public static final int DELAY = 2;
 	public static final MapCodec<ShulkerTendonBlock> CODEC = simpleCodec(ShulkerTendonBlock::new);
 
 	@Override
@@ -53,6 +61,22 @@ public class ShulkerTendonBlock extends DirectionalBlock {
 	}
 
 	@Override
+	@Nullable
+	public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+		return new ShulkerTendonBlockEntity(pos, state);
+	}
+
+	@Override
+	@Nullable
+	public <T extends BlockEntity> BlockEntityTicker<T> getTicker(
+		Level level,
+		BlockState state,
+		BlockEntityType<T> blockEntityType
+	) {
+		return ShulkerTendonBlockEntity::tick;
+	}
+
+	@Override
 	protected void neighborChanged(
 		BlockState state,
 		Level level,
@@ -75,17 +99,32 @@ public class ShulkerTendonBlock extends DirectionalBlock {
 		var targetState = level.getBlockState(targetPos);
 		var targetStrength = targetState.getDestroySpeed(level, targetPos);
 
+		if (level.getBlockEntity(pos) instanceof ShulkerTendonBlockEntity be && be.getCooldown() > 0) {
+			return;
+		}
+
 		if (targetStrength > 5 || targetStrength < 0 || !level.isInWorldBounds(targetPos)) {
 			return;
 		}	
 
-		// TODO: make this work
+		// TODO: make sounds work
 		level.playLocalSound(pos, SoundEvents.SHULKER_TELEPORT, SoundSource.BLOCKS, 5, 1, false);
 
 		level.destroyBlock(targetPos, true);
+		level.setBlock(pos, Blocks.AIR.defaultBlockState(), UPDATE_ALL);
 
 		level.setBlock(targetPos, state, UPDATE_ALL);
-		level.setBlock(pos, Blocks.AIR.defaultBlockState(), UPDATE_ALL);
+		if (level.getBlockEntity(targetPos) instanceof ShulkerTendonBlockEntity be) {
+			be.setCooldown((targetStrength * 8) + 2);
+		} else {
+			Recharged.LOGGER.error("Failed to set cooldown for new Shulker Tendon");
+		}
+
+		for (var entity : level.getEntities(null, new AABB(targetPos))) {
+			if (entity instanceof LivingEntity le) {
+				le.hurt(level.damageSources().generic(), 8);
+			}
+		}
 	}
 
 	@Override
